requirements.txt
streamlit
google-genai
requests
beautifulsoup4
newspaper3k
lxml_html_clean
python-dotenv
lxml
